had to update jenkins and moved snapshot to ComputeEngineComputer. Also, added only for failed jobs
